refactor(BA-6669): move agent status/permission defs down from models to data
Relocate the pure-enum rbac permission_defs module from
models/rbac/permission_defs.py to data/permission/permission_defs.py and
redirect data/agent/modifier.py to import AgentStatus (already defined in
data/agent/types.py) directly. All models/repositories/api importers of
the permission enums now point at the data layer. This makes data/agent a
pure leaf and removes 2 more data->models back-edges feeding the
models<->data cycle. The enum definitions themselves are unchanged.
