+++ src/ai/backend/manager/api/gql_legacy/domain.py
@@ -26,6 +26,10 @@ from ai.backend.manager.data.domain.types import (
     DomainData,
     UserInfo,
 )
+from ai.backend.manager.data.permission.permission_defs import (
+    DomainPermission,
+    ScalingGroupPermission,
+)
 from ai.backend.manager.models.domain import DomainRow, domains, get_permission_ctx
 from ai.backend.manager.models.minilang import FieldSpecItem, OrderSpecItem
 from ai.backend.manager.models.minilang.ordering import QueryOrderParser
@@ -35,7 +39,6 @@ from ai.backend.manager.models.rbac import (
     SystemScope,
 )
 from ai.backend.manager.models.rbac.context import ClientContext
-from ai.backend.manager.models.rbac.permission_defs import DomainPermission, ScalingGroupPermission
 from ai.backend.manager.models.scaling_group import get_scaling_groups
 from ai.backend.manager.models.user import UserRole
 from ai.backend.manager.repositories.base.creator import Creator
