+++ src/ai/backend/manager/data/permission/permission_defs.py
