+++ src/ai/backend/manager/models/group/row.py
@@ -34,6 +34,7 @@ from ai.backend.common import msgpack
 from ai.backend.common.types import ResourceSlot, VFolderHostPermissionMap
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.data.group.types import GroupData, ProjectType
+from ai.backend.manager.data.permission.permission_defs import ProjectPermission
 from ai.backend.manager.defs import RESERVED_DOTFILES
 from ai.backend.manager.errors.common import ObjectNotFound
 from ai.backend.manager.models.association_container_registries_groups import (
@@ -60,7 +61,6 @@ from ai.backend.manager.models.rbac import (
     required_permission,
 )
 from ai.backend.manager.models.rbac.context import ClientContext
-from ai.backend.manager.models.rbac.permission_defs import ProjectPermission
 from ai.backend.manager.models.types import (
     QueryCondition,
     QueryOption,
